requirements.txt
torch
nltk
numpy
mysql-connector-python
fuzzywuzzy
python-Levenshtein
